# openapi.yaml
openapi: 3.1.0
info:
  title: Human Design API
  version: 4.0.0
paths:
...
  /transits/solar_return:
    get:
      summary: Get Solar Return
      operationId: get_solar_return_transits_solar_return_get
      security:
      - HTTPBearer: []
      parameters:
      - name: year
        in: query
        required: true
        schema:
          type: integer
          description: Birth year
          title: Year
        description: Birth year
      - name: month
        in: query
        required: true
        schema:
          type: integer
          description: Birth month
          title: Month
        description: Birth month
      - name: day
        in: query
        required: true
        schema:
          type: integer
          description: Birth day
          title: Day
        description: Birth day
      - name: hour
        in: query
        required: true
        schema:
          type: integer
          description: Birth hour
          title: Hour
        description: Birth hour
      - name: minute
        in: query
        required: false
        schema:
          type: integer
          description: Birth minute (default 0)
          default: 0
          title: Minute
        description: Birth minute (default 0)
      - name: second
        in: query
        required: false
        schema:
          type: integer
          description: Birth second (optional, default 0)
          default: 0
          title: Second
        description: Birth second (optional, default 0)
      - name: place
        in: query
        required: true
        schema:
          type: string
          description: Birth place (city, country)
          title: Place
        description: Birth place (city, country)
      - name: sr_year_offset
        in: query
        required: false
        schema:
          type: integer
          description: Calculate Solar Return for X years after birth. 0 is the current SR.
          default: 0
          title: Sr Year Offset
        description: Calculate Solar Return for X years after birth. 0 is the current SR.
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /transits/daily:
    get:
      summary: Get Daily Transit
      operationId: get_daily_transit_transits_daily_get
      security:
      - HTTPBearer: []
      parameters:
      - name: year
        in: query
        required: true
        schema:
          type: integer
          description: Birth year
          title: Year
        description: Birth year
      - name: month
        in: query
        required: true
        schema:
          type: integer
          description: Birth month
          title: Month
        description: Birth month
      - name: day
        in: query
        required: true
        schema:
          type: integer
          description: Birth day
          title: Day
        description: Birth day
      - name: hour
        in: query
        required: true
        schema:
          type: integer
          description: Birth hour
          title: Hour
        description: Birth hour
      - name: minute
        in: query
        required: false
        schema:
          type: integer
          description: Birth minute (default 0)
          default: 0
          title: Minute
        description: Birth minute (default 0)
      - name: second
        in: query
        required: false
        schema:
          type: integer
          description: Birth second (optional, default 0)
          default: 0
          title: Second
        description: Birth second (optional, default 0)
      - name: place
        in: query
        required: true
        schema:
          type: string
          description: Birth place (city, country)
          title: Place
        description: Birth place (city, country)
      - name: transit_year
        in: query
        required: true
        schema:
          type: integer
          description: Transit year to analyze
          title: Transit Year
        description: Transit year to analyze
      - name: transit_month
        in: query
        required: true
        schema:
          type: integer
          description: Transit month to analyze
          title: Transit Month
        description: Transit month to analyze
      - name: transit_day
        in: query
        required: true
        schema:
          type: integer
          description: Transit day to analyze
          title: Transit Day
        description: Transit day to analyze
      - name: current_place
        in: query
        required: false
        schema:
          type: string
          description: Current location for transit calculation (optional, defaults to Birth Place)
          title: Current Place
        description: Current location for transit calculation (optional, defaults to Birth Place)
      - name: transit_hour
        in: query
        required: false
        schema:
          type: integer
          description: Transit hour (local time at calculation place, default 12)
          default: 12
          title: Transit Hour
        description: Transit hour (local time at calculation place, default 12)
      - name: transit_minute
        in: query
        required: false
        schema:
          type: integer
          description: Transit minute (default 0)
          default: 0
          title: Transit Minute
        description: Transit minute (default 0)
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta:
                    type: object
                    properties:
                      transit_date_local:
                        type: string
                      transit_date_utc:
                        type: string
                      birth_date_utc:
                        type: string
                      birth_date:
                        type: string
                        description: "Birth date in ISO format"
                      create_date:
                        type: string
                        description: "Design/Creation date"
                      place:
                        type: string
                        description: "Birth place"
                      age:
                        type: integer
                      gender:
                        type: string
                      islive:
                        type: boolean
                      zodiac_sign:
                        type: string
                      energy_type:
                        type: string
                      strategy:
                        type: string
                      signature:
                        type: string
                      not_self:
                        type: string
                      aura:
                        type: string
                      inner_authority:
                        type: string
                      inc_cross:
                        type: string
                      profile:
                        type: string
                      calculation_place:
                        type: string
                      location:
                        type: string
                      type:
                        type: string
                      authority:
                        type: string
                      total_centers:
                        type: integer
                      defined_centers:
                        type: array
                        items:
                          type: string
                      undefined_centers:
                        type: array
                        items:
                          type: string
                      definition:
                        type: string
                      channels:
                        type: object
                        properties:
                          Channels:
                            type: array
                            items:
                              type: object
                              properties:
                                channel:
                                  type: string
                                  description: "Formatted channel string (e.g. '6/59: Name (Desc)')"
                  composite_changes:
                    type: object
                    properties:
                      new_channels:
                        type: array
                        items:
                          type: object
                          properties:
                            gates:
                              type: string
                            name:
                              type: string
                            description:
                              type: string
                      new_centers:
                        type: array
                        items:
                          type: string
                  planetary_transits:
                    type: array
                    items:
                      type: object
                      properties:
                        planets:
                          type: string
                        gate:
                          type: integer
                        line:
                          type: integer
                        color:
                          type: integer
                        tone:
                          type: integer
                        base:
                          type: integer
                        lon:
                          type: number
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer